showing Connections on its own tab
