+++ app/react/Viewer/actions/specs/uiActions.spec.js
@@ -58,13 +58,13 @@ describe('Viewer uiActions', () => {
     it('should dispatch a ACTIVATE_REFERENCE with id', () => {
       actions.activateReference('id')(dispatch);
       expect(dispatch).toHaveBeenCalledWith({type: types.ACTIVE_REFERENCE, reference: 'id'});
-      expect(dispatch).toHaveBeenCalledWith({type: types.OPEN_PANEL, panel: 'viewReferencesPanel'});
+      expect(dispatch).toHaveBeenCalledWith({type: types.OPEN_PANEL, panel: 'viewMetadataPanel'});
     });
 
     it('should scroll to the elements', () => {
       actions.activateReference('id')(dispatch);
       expect(scroller.to).toHaveBeenCalledWith('.document-viewer a[data-id="id"]', '.document-viewer');
-      expect(scroller.to).toHaveBeenCalledWith('.document-references .item[data-id="id"]', '.document-references .sidepanel-body');
+      expect(scroller.to).toHaveBeenCalledWith('.document-metadata .item[data-id="id"]', '.document-metadata .sidepanel-body');
     });
   });
 
@@ -80,7 +80,7 @@ describe('Viewer uiActions', () => {
 
     it('should dispatch a call to activateReference', () => {
       expect(dispatch).toHaveBeenCalledWith({type: types.ACTIVE_REFERENCE, reference: 'id2'});
-      expect(dispatch).toHaveBeenCalledWith({type: types.OPEN_PANEL, panel: 'viewReferencesPanel'});
+      expect(dispatch).toHaveBeenCalledWith({type: types.OPEN_PANEL, panel: 'viewMetadataPanel'});
     });
 
     it('should dispatch a SET_TARGET_SELECTION with found range', () => {
