+++ app/react/Viewer/components/ViewMetadataPanel.js
@@ -10,6 +10,8 @@ import {actions as formActions} from 'react-redux-form';
 import DocumentForm from '../containers/DocumentForm';
 import {ShowDocument} from 'app/Documents';
 import modals from 'app/Modals';
+import {Tabs, TabLink, TabContent} from 'react-tabs-redux';
+import Connections from './ConnectionsList';
 
 export class ViewMetadataPanel extends Component {
   close() {
@@ -28,18 +30,39 @@ export class ViewMetadataPanel extends Component {
     const {doc, docBeingEdited} = this.props;
 
     return (
-      <SidePanel open={this.props.open}>
+      <SidePanel open={this.props.open} className="document-metadata">
         <div className="sidepanel-header">
           <h1>Metadata</h1>
           <i className="fa fa-close close-modal" onClick={this.close.bind(this)}/>
         </div>
         <div className="sidepanel-body">
-          {(() => {
-            if (docBeingEdited) {
-              return <DocumentForm onSubmit={this.submit.bind(this)} />;
-            }
-            return <ShowDocument doc={doc}/>;
-          })()}
+          <Tabs>
+            <ul className="nav nav-tabs">
+              <li>
+                <TabLink to="toc">Table of contents</TabLink>
+              </li>
+              <li>
+                <TabLink to="metadata" default>Metadata</TabLink>
+              </li>
+              <li>
+                <TabLink to="connections">Connections (22)</TabLink>
+              </li>
+            </ul>
+            <TabContent for="toc">
+              TOC
+            </TabContent>
+            <TabContent for="metadata">
+              {(() => {
+                if (docBeingEdited) {
+                  return <DocumentForm onSubmit={this.submit.bind(this)} />;
+                }
+                return <ShowDocument doc={doc}/>;
+              })()}
+            </TabContent>
+            <TabContent for="connections">
+              <Connections />
+            </TabContent>
+          </Tabs>
         </div>
       </SidePanel>
     );
