+++ app/react/Viewer/components/specs/ConnectionsList.spec.js
@@ -3,10 +3,10 @@ import {shallow} from 'enzyme';
 import configureMockStore from 'redux-mock-store';
 import Immutable from 'immutable';
 
-import PanelContainer, {ViewReferencesPanel} from 'app/Viewer/components/ViewReferencesPanel';
+import ListContainer, {ConnectionsList} from 'app/Viewer/components/ConnectionsList';
 import SidePanel from 'app/Layout/SidePanel';
 
-describe('ViewReferencesPanel', () => {
+describe('ConnectionsList', () => {
   let component;
   let props;
 
@@ -31,15 +31,8 @@ describe('ViewReferencesPanel', () => {
   });
 
   let render = () => {
-    component = shallow(<ViewReferencesPanel {...props}/>);
+    component = shallow(<ConnectionsList {...props}/>);
   };
-
-  it('should render a SidePanel', () => {
-    render();
-
-    expect(component.find(SidePanel).length).toBe(1);
-    expect(component.find(SidePanel).props().open).toBe(false);
-  });
 
   it('should merge and render references in order with the proper document titles', () => {
     render();
@@ -87,7 +80,7 @@ describe('ViewReferencesPanel', () => {
     beforeEach(() => {
       props.uiState = Immutable.fromJS({
         reference: Immutable.fromJS({targetRange: 'targetRange'}),
-        panel: 'viewReferencesPanel',
+        panel: 'ConnectionsList',
         activeReference: 'ref1'
       });
     });
@@ -121,16 +114,7 @@ describe('ViewReferencesPanel', () => {
     });
   });
 
-  describe('when props.referencePanel', () => {
-    it('should open SidePanel', () => {
-      props.uiState = Immutable.fromJS({panel: 'viewReferencesPanel'});
-      render();
-
-      expect(component.find(SidePanel).props().open).toBe(true);
-    });
-  });
-
-  describe('PanelContainer', () => {
+  describe('ListContainer', () => {
     let state = {
       documentViewer: {
         uiState: Immutable.fromJS({
@@ -145,7 +129,7 @@ describe('ViewReferencesPanel', () => {
 
     let renderContainer = () => {
       let store = mockStore(state);
-      component = shallow(<PanelContainer />, {context: {store}});
+      component = shallow(<ListContainer />, {context: {store}});
     };
 
     it('should should map props', () => {
