+++ package.json
@@ -72,6 +72,7 @@
     "react-redux": "^4.0.0",
     "react-redux-form": "0.13.9",
     "react-router": "2.0.0",
+    "react-tabs-redux": "^1.3.0",
     "read-multiple-files": "^1.1.1",
     "redux": "^3.0.4",
     "redux-thunk": "^2.0.1",
